Feat: [BADA-321] 가맹점 목록 정렬 모달 UI 거래 파트와 통일 #217
